# Development of Web Applications/Project/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Battleship Application API
  description: |-
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
paths:
...
  /user:
    get:
      tags:
        - user
      summary: gets user info
      responses:
        '200':
          description: successful operation
          content:
            application/vnd.siren+json:
              schema:
                $ref: '#/components/schemas/UserHomeOutputModel'
    post:
      tags:
        - user
      summary: Create user
      description: This can be used by anyone who wishes to create a user for themselves.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/vnd.siren+json:
            schema:
              $ref: '#/components/schemas/UserCreateInputModel'
      responses:
        '201':
          description: successful operation
          content:
            application/vnd.siren+json:
              schema:
                $ref: '#/components/schemas/CreateUser'
        '400':
          description: insecure password
        '409':
          description: user already exists
...
  /user/signIn:
    post:
      tags:
        - user
      summary: Creates a token for a user
      description: Creates a token for the user, after confirming username and password passes through the request body
      operationId: createToken
      requestBody:
        content:
          application/vnd.siren+json:
            schema:
              $ref: '#/components/schemas/UserCreateInputModel'
      responses:
        '200':
          description: Token created
          content:
            application/vnd.siren+json:
              schema:
                $ref: '#/components/schemas/tokenOutput'
        '401':
          description: authentication failed
...
  /user/leaderboard:
    get:
      tags:
        - user
      summary: Gets the leaderboard
      description: Gets the leaderboard without the need of authorization, order by either games played or games won
      operationId: getLeaderboard
      parameters:
        - in: header
          name: orderBy
          schema:
            type: string
            example: "Wins"
        - in: header
          name: top
          schema:
            type: number
            example: 10
        - in: header
          name: idx
          schema:
            type: number
            example: 10
        - in: header
          name: name
          schema:
            type: string
            example: "username"
      responses:
        '201':
          description: leaderboard found
          content:
            application/vnd.siren+json:
              schema:
                $ref: '#/components/schemas/leaderboard'
        '404':
          description: order by invalid
  /user/{id}:
    get:
      tags:
        - user
      summary: get user info by id
      parameters:
        - in: path
          name: id
          schema:
            type: string
            example: "userID"
          required: true
      responses:
        '201':
          description: user found
          content:
            application/vnd.siren+json:
              schema:
                $ref: '#/components/schemas/UserHomeOutputModel'
        '401':
          description: authentication failed
        '404':
          description: user not found
...
components:
  schemas:
...
    CreateUser:
      type: object
      properties:
        clazz:
          type: array
          items:
            type: string
          example: [ "UserHomeOutputModel" ]
        properties:
          type: object
          items:
            $ref: '#/components/schemas/UserHomeOutputModel'
        siren:
          type: object
          items:
            $ref: '#/components/schemas/siren'
    UserHomeOutputModel:
      type: object
      properties:
        id:
          type: string
          example: "101"
        username:
          type: string
          example: "username"
        stats:
          type: number
          example: 1
...
    leaderboard:
      type: object
      properties:
        clazz:
          type: array
          items:
            type: string
          example: [ "LeaderboardOutputModel" ]
        properties:
          type: object
          properties:
            stats:
              type: array
              items:
                $ref: '#/components/schemas/stats'
        siren:
          type: object
          items:
            $ref: '#/components/schemas/siren'
    stats:
      type: object
      properties:
            id:
              type: number
              example: 1
            username:
              type: string
              example: "usernameHere"
            gamesPlayed:
              type: number
              example: 10
            gamesWon:
              type: number
              example: 5
...
    UserCreateInputModel:
      type: object
      properties:
        username:
          type: string
          example: theUser
        password:
          type: string
          example: 'Th1s1sAP4ssword'
    tokenOutput:
      type: object
      properties:
        clazz:
          type: array
          items:
            type: string
          example: [ "UserTokenCreateOutputModel" ]
        properties:
          type: object
          properties:
            token:
              type: string
              example: "weptgnejtnwpntowet"
        siren:
          type: object
          items:
            $ref: '#/components/schemas/siren'
...
    siren:
      type: object
      properties:
        links:
            $ref: '#/components/schemas/Rels'
        entities:
            $ref: '#/components/schemas/Entities'
        actions:
            $ref: '#/components/schemas/Actions'
    Entities:
      type: object
      properties:
        properties:
          type: object
        links:
          type: array
          items:
            $ref: '#/components/schemas/Rels'
    Actions:
      type: object
    Rels:
      type: object
      properties:
        rel:
          type: array
          items:
            type: string
          example: ["self"]
        href:
          type: string
          example: "/user/eeefdd87-ec51-4fb4-b443-03e5f63c3611"
...